organization module changes

+++ src/modules/organization/user/users.controller.ts
@@ -12,20 +12,28 @@ import {
 } from '@nestjs/common';
 import { findUsersFilters } from './dtos/find-users-filter';
 import { UpdateUserBody } from './dtos/update-user-body';
-import { UserRepository } from '../../../repositories/user-repository';
-import { AccessValidator } from 'src/modules/auth/decorators/access-validator.decorator';
+import { UserRepository } from 'src/repositories/user-repository';
 import { actions } from 'src/services/permissions/permissions';
+import { AccessValidator } from 'src/modules/auth/decorators/access-validator.decorator';
 @Controller('organizations/users')
 export class UsersController {
-  constructor(private userRepository: UserRepository) {}
+  private validateAccess: (action: actions) => void;
+
+  constructor(private userRepository: UserRepository) {
+    this.validateAccess = AccessValidator;
+  }
 
   @Get()
   async findAll(@Body() body: findUsersFilters): Promise<User[]> {
+    this.validateAccess('findAllUsers');
+
     return this.userRepository.findAll(body);
   }
 
   @Get('/:id')
   async findById(@Param('id') id: string, @Res() res: any) {
+    this.validateAccess('findUserById');
+
     const user = this.userRepository.findById(id);
 
     const status = user ? 200 : 404;
@@ -33,12 +41,9 @@ export class UsersController {
   }
 
   @Post()
-  async create(
-    @AccessValidator() validateUserAccess: (action: actions) => void,
-    @Body() body: CreateUserBody,
-    @Res() res: any,
-  ) {
-    validateUserAccess('createUser');
+  async create(@Body() body: CreateUserBody, @Res() res: any) {
+    this.validateAccess('createUser');
+
     const {
       name,
       last_name,
@@ -70,6 +75,8 @@ export class UsersController {
     @Param('id') id: string,
     @Res() res: any,
   ) {
+    this.validateAccess('updateUser');
+
     const {
       name,
       last_name,
@@ -99,6 +106,8 @@ export class UsersController {
 
   @Delete('/:id')
   async delete(@Param('id') id: string, @Res() res: any) {
+    this.validateAccess('deleteUser');
+
     const user = await this.userRepository.delete(id);
 
     const status = user ? 204 : 404;
